+++ src/v3/containers/Portfolio/Portfolio.test.js
@@ -27,6 +27,7 @@ jest.mock("./About", () => () => <section>About content</section>);
 jest.mock("./Skills", () => () => <section>Skills content</section>);
 jest.mock("./Experience", () => () => <section>Experience content</section>);
 jest.mock("./Projects", () => () => <section>Projects content</section>);
+jest.mock("./Blog", () => () => <section>Blog content</section>);
 jest.mock("./Contact", () => () => <section>Contact content</section>);
 jest.mock("./Quote", () => () => <section>Quote content</section>);